Switch to using maze.runCode
Last year's iteration used a hacky way of getting the code to run on the maze - it defined some functions in the global scope and relied on calling those.

This commit introduces maze.runCode, which defines a function dynamically based on the player's code and binds the maze's functions as arguments so that they are in scope.